1892: Move activity log to the user menu
