+++ administration/src/bp-modules/user-settings/UserSettingsController.tsx
@@ -4,7 +4,6 @@ import styled from 'styled-components'
 import { useWhoAmI } from '../../WhoAmIProvider'
 import { Role } from '../../generated/graphql'
 import { ProjectConfigContext } from '../../project-configs/ProjectConfigContext'
-import ActivityLogCard from './ActivityLogCard'
 import ChangePasswordForm from './ChangePasswordForm'
 import NotificationSettings from './NotificationSettings'
 
@@ -17,13 +16,12 @@ const UserSettingsContainer = styled.div`
 `
 
 const UserSettingsController = (): ReactElement => {
-  const { applicationFeature, activityLogConfig, projectId } = useContext(ProjectConfigContext)
+  const { applicationFeature, projectId } = useContext(ProjectConfigContext)
   const { role } = useWhoAmI().me
   return (
     <UserSettingsContainer>
       {applicationFeature && role !== Role.ProjectAdmin && <NotificationSettings projectId={projectId} />}
       <ChangePasswordForm />
-      {activityLogConfig && <ActivityLogCard activityLogConfig={activityLogConfig} />}
     </UserSettingsContainer>
   )
 }
