+++ administration/src/util/translations/de.json
@@ -281,6 +281,7 @@
     "showPassword": "Passwort anzeigen",
     "hidePassword": "Passwort verstecken",
     "userSettings": "Benutzereinstellungen",
+    "activityLog": "Aktivitätsprotokoll",
     "logout": "Logout",
     "confirm": "Bestätigen",
     "delete": "Löschen",
@@ -437,10 +438,6 @@
     "csvImportUnknownError": "Beim Verarbeiten der Datei ist ein unbekannter Fehler aufgetreten"
   },
   "userSettings": {
-    "activityLog": "Aktivitätsprotokoll",
-    "viewActivity": "Aktivitäten ansehen",
-    "activityLogDescription": "Hier können sie den Verlauf ihrer Aktivitäten einsehen.",
-    "noEntries": "Keine Einträge vorhanden",
     "passwordChangeSuccessful": "Passwort erfolgreich geändert.",
     "changePassword": "Passwort ändern",
     "changePasswordExplanation": "Ein gültiges Passwort ist mindestens zwölf Zeichen lang, enthält mindestens einen Klein- und einen Großbuchstaben sowie mindestens ein Sonderzeichen.",
@@ -453,6 +450,11 @@
     "newApplications": "Neue Anträge",
     "newVerifications": "Antragsverifizierungen",
     "save": "Speichern"
+  },
+  "activityLog": {
+    "viewActivity": "Aktivitäten ansehen",
+    "activityLogDescription": "Hier können Sie den Verlauf der erstellten Karten einsehen. Bitte beachten Sie, dass diese Informationen nur innerhalb der aktuellen Sitzung verfügbar sind - sobald Sie den Tab schließen, werden die Daten gelöscht.",
+    "noEntries": "Keine Einträge vorhanden"
   },
   "users": {
     "addUserSuccess": "Benutzer erfolgreich erstellt.",
